+++ tests/fixtures/Attribute/ClassWithStringPropertyBeforeSet.php
@@ -6,7 +6,7 @@ namespace Membrane\Fixtures\Attribute;
 
 use Membrane\Attribute\Placement;
 use Membrane\Attribute\SetFilterOrValidator;
-use Membrane\Validator\Object\RequiredFields;
+use Membrane\Validator\FieldSet\RequiredFields;
 
 #[SetFilterOrValidator(new RequiredFields('property'), Placement::BEFORE)]
 class ClassWithStringPropertyBeforeSet
