+++ tests/fixtures/Attribute/Docs/BlogPostMaxTags.php
@@ -9,8 +9,8 @@ use Membrane\Attribute\Placement;
 use Membrane\Attribute\SetFilterOrValidator;
 use Membrane\Attribute\Subtype;
 use Membrane\Filter\Type\ToString;
-use Membrane\Validator\Array\Count;
-use Membrane\Validator\Object\RequiredFields;
+use Membrane\Validator\Collection\Count;
+use Membrane\Validator\FieldSet\RequiredFields;
 
 
 #[SetFilterOrValidator(new RequiredFields('title', 'body', 'tags'), Placement::BEFORE)]
