+++ tests/fixtures/Attribute/Docs/BlogPostRegexAndMaxLength.php
@@ -9,8 +9,8 @@ use Membrane\Attribute\Placement;
 use Membrane\Attribute\SetFilterOrValidator;
 use Membrane\Attribute\Subtype;
 use Membrane\Filter\Type\ToString;
-use Membrane\Validator\Array\Count;
-use Membrane\Validator\Object\RequiredFields;
+use Membrane\Validator\Collection\Count;
+use Membrane\Validator\FieldSet\RequiredFields;
 use Membrane\Validator\String\Length;
 use Membrane\Validator\String\Regex;
