+++ tests/fixtures/Attribute/Docs/BlogPostWithAllOf.php
@@ -9,8 +9,8 @@ use Membrane\Attribute\Placement;
 use Membrane\Attribute\SetFilterOrValidator;
 use Membrane\Attribute\Subtype;
 use Membrane\Filter\Type\ToString;
-use Membrane\Validator\Array\Count;
-use Membrane\Validator\Object\RequiredFields;
+use Membrane\Validator\Collection\Count;
+use Membrane\Validator\FieldSet\RequiredFields;
 use Membrane\Validator\String\Length;
 use Membrane\Validator\String\Regex;
 use Membrane\Validator\Utility\AllOf;
@@ -20,7 +20,7 @@ class BlogPostWithAllOf
 {
     public function __construct(
         #[FilterOrValidator(new ToString())]
-        #[FilterOrValidator(new AllOf(new Length(5,50), new Regex('#^([A-Z][a-z]*\s){0,9}([A-Z][a-z]*)$#')))]
+        #[FilterOrValidator(new AllOf(new Length(5, 50), new Regex('#^([A-Z][a-z]*\s){0,9}([A-Z][a-z]*)$#')))]
         public string $title,
         #[FilterOrValidator(new ToString())]
         public string $body,
